+++ tests/notifications/conftest.py
@@ -1,0 +1,7 @@
+from pytest_factoryboy import register
+
+from meinberlin.test.factories import budgeting as budgeting_factories
+from meinberlin.test.factories import offlineevents
+
+register(budgeting_factories.ProposalFactory)
+register(offlineevents.OfflineEventFactory)